Bug 26576: Subfield descriptions on authority detail view are cut off
This patch corrects the authority detail view on the staff interface so
that subfield descriptions are not longer output as truncated strings.

For instance, before this patch the subfield description for 035$a would
be "System control numbe"

After the patch it should be "System control number."

Based on the assumption that the subfield descriptions were truncated at
some point for readability reasons I've also made some style changes
intended to improve the readability of the page.

To test, apply the patch and go to Authorities.

- Search for and view the details for an authority.
- On the authority detail page, confirm that subfield descriptions are
  shown in full.
- Confirm that the style changes look good and make the data more
  readable.

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>