Bug 26729: Add focus class to vendor name field in vendor add form
This patch adds the "focus" class to the vendor name field in the form
for adding or editing a vendor in acquisitions.

To test, apply the patch and go to Acquisitions -> New vendor.

On page load the cursor focus should automatically be on the vendor name
field. The same should true when editing an existing vendor.

Signed-off-by: Séverine QUEUNE <severine.queune@bulac.fr>

Signed-off-by: Barbara Johnson <barbara.johnson@bedfordtx.gov>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>