Bug 26763: Use standard information style for multi-hold message
This patch modifies the markup of the message which is displayed when
the patron tries to place a hold on a title they already have a hold on.
Now the message is styled consistently with other similar messages.

To test, apply the patch and follow the steps used to test Bug 19718:

- In Administration -> Circulation and fine rules, set a rule to allow
  multiple holds on a title.
- Log in to the OPAC as a user covered by the above rule and locate
  a bibliographic record on which multiple holds can be placed.
- Place a hold on the record.
- Return to the record details for that title and place another hold.
- You should see a correctly-styled message about having at least one
  other hold on the record.

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>